переносит код в модули
